backend/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.0
backcall==0.2.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator
Django
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.14.0
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna
ipython
gunicorn
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
protobuf==4.21.10
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.6.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
django-cors-headers